+++ plugins/Passbolt/WebInstaller/src/Utility/PassboltConfiguration.php
@@ -19,7 +19,6 @@ use Cake\View\ViewVarsTrait;
 
 class PassboltConfiguration
 {
-
     use ViewVarsTrait;
 
     /**