+++ tests/TestCase/Model/Table/Comments/FindViewForeignCommentsTest.php
@@ -24,6 +24,7 @@ use Cake\ORM\TableRegistry;
 class FindViewForeignCommentsTest extends AppTestCase
 {
     use CommentsModelTrait;
+
     public $Comments;
 
     public $fixtures = ['app.Base/Resources', 'app.Base/Users', 'app.Base/Profiles', 'app.Base/Groups', 'app.Base/GroupsUsers', 'app.Base/Permissions', 'app.Base/Comments', 'app.Base/Avatars'];